Explorer-style table widget with dynamic columns
Rework the dashboard table widget to show raw CUR line items (like the
Explorer) instead of aggregated cost rows.

- Extract DataTable, ColumnsPicker, and cell rendering from explorer.tsx
  into a shared components/data-table.tsx
- Table widget uses queryExplorerRows to fetch raw rows filtered by the
  view's global filters and date range
- Full column picker with drag-to-reorder, show/hide, tag columns
- Sortable column headers
- Table WidgetSpec no longer needs groupBy — stores hiddenColumns and
  columnOrder instead
- Remove TableColumn enum — columns are dynamic, driven by the dataset
- Update validator, serializer, seed views, and widget inspector
