+++ packages/core/src/types/seed-views.ts
@@ -29,9 +29,6 @@ export const OVERVIEW_SEED_VIEW: ViewSpec = {
           id: 'overview-breakdown',
           type: 'table',
           size: 'full',
-          groupBy: asDimensionId('account'),
-          columns: ['entity', 'cost', 'delta', 'percentChange'],
-          topN: 20,
         },
       ],
     },
